improved maintainability for managers
